backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-dotenv
phonenumbers
passlib[bcrypt]
python-jose[cryptography]
httpx
aiobotocore
python-magic
slowapi


fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-dotenv
pydantic-settings
phonenumbers
passlib[bcrypt]
python-jose[cryptography]
httpx
aiobotocore
python-magic
slowapi